Add basic feature scaffold with router and models
