+++ .env.example
@@ -1,0 +1,7 @@
+FIREBASE_PROJECT_ID=operationsummer-prod
+GOOGLE_CLIENT_ID_IOS=
+GOOGLE_CLIENT_ID_ANDROID=
+GOOGLE_CLIENT_ID_WEB=
+SHEETS_SPREADSHEET_ID=
+DEFAULT_PRINTER=HomePrinter
+SERVICE_ACCOUNT_JSON={...}
