+++ .github/workflows/flutter-ci.yml
@@ -1,0 +1,14 @@
+name: Flutter CI
+on:
+  push:
+    branches: [ main ]
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+    - uses: actions/checkout@v4
+    - uses: subosito/flutter-action@v2
+      with: { flutter-version: '3.22.0' }
+    - run: flutter pub get -C app
+    - run: flutter analyze -C app
+    - run: flutter test -C app
