+++ .gitignore
@@ -1,0 +1,11 @@
+# Flutter / Dart
+.dart_tool/
+.packages
+.pub/
+build/
+# Node
+functions/node_modules/
+print-daemon/node_modules/
+# IDE
+.idea/
+.vscode/
