+++ README.md
@@ -1,2 +1,5 @@
-# OperationSummer
+# Operation Summer
+
+Initial scaffolding auto-generated. Run `scripts/dev_setup.sh` after cloning.
+
 A family management app to assist with schedules, chores, entertainment, and education.
