+++ app/lib/features/calendar/calendar_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class CalendarScreen extends StatelessWidget {
+  const CalendarScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Calendar')),
+      body: const Center(child: Text('Calendar coming soon')),
+    );
+  }
+}
