+++ app/lib/features/children/providers/children_provider.dart
@@ -1,0 +1,4 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import '../models/child.dart';
+
+final childrenProvider = StateProvider<List<Child>>((ref) => []);
