+++ app/lib/features/chores/providers/chores_provider.dart
@@ -1,0 +1,4 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import '../models/chore.dart';
+
+final choresProvider = StateProvider<List<Chore>>((ref) => []);
