+++ app/lib/features/dashboard/dashboard_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class DashboardScreen extends StatelessWidget {
+  const DashboardScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Mission Dashboard')),
+      body: const Center(child: Text('Dashboard coming soon')),
+    );
+  }
+}
