+++ app/lib/features/learning/learning_hub_screen.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class LearningHubScreen extends StatelessWidget {
+  const LearningHubScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(title: const Text('Learning Hub')),
+      body: const Center(child: Text('Learning content coming soon')),
+    );
+  }
+}
