+++ app/lib/features/learning/providers/lessons_provider.dart
@@ -1,0 +1,4 @@
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import '../models/lesson.dart';
+
+final lessonsProvider = StateProvider<List<Lesson>>((ref) => []);
