+++ app/pubspec.yaml
@@ -36,6 +36,8 @@ dependencies:
   cupertino_icons: ^1.0.8
   lottie: ^3.3.1
   google_fonts: ^6.2.1
+  flutter_riverpod: ^2.4.0
+  go_router: ^13.2.0
 
 dev_dependencies:
   flutter_test:
