+++ firestore.indexes.json
@@ -1,0 +1,3 @@
+{
+  "indexes": []
+}
