+++ functions/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "operationsummer-functions",
+  "private": true,
+  "engines": { "node": "20" },
+  "scripts": {
+    "build": "tsc",
+    "deploy": "firebase deploy --only functions"
+  },
+  "dependencies": {
+    "firebase-admin": "^12.0.0",
+    "firebase-functions": "^4.0.0",
+    "googleapis": "^133.0.0",
+    "pdfkit": "^0.14.0"
+  },
+  "devDependencies": {
+    "typescript": "^5.4.0"
+  }
+}
