+++ functions/src/sheets.sync.ts
@@ -1,0 +1,1 @@
+// TODO: parse webhook events and update Firestore accordingly
