+++ functions/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "compilerOptions": {
+    "lib": ["es2021"],
+    "module": "commonjs",
+    "target": "es2021",
+    "outDir": "lib",
+    "esModuleInterop": true,
+    "strict": true
+  },
+  "include": ["src"]
+}
