+++ scripts/dev_setup.sh
@@ -1,0 +1,5 @@
+#!/usr/bin/env bash
+set -e
+flutter pub get -C app
+npm ci -C functions
+npm ci -C print-daemon
